+++ src/pint/models/stand_alone_psr_binaries/ELL1H_model.py
@@ -3,22 +3,36 @@
 # This Python file uses the following encoding: utf-8
 import astropy.units as u
 import numpy as np
+from loguru import logger as log
 
 from .ELL1_model import ELL1BaseModel
 
 
 class ELL1Hmodel(ELL1BaseModel):
-    """ELL1H pulsar binary model using H3, H4 or STIGMA as shapiro delay parameter.
+    """ELL1H pulsar binary model using H3, H4 or STIGMA as shapiro delay parameters.
 
     Note
     ----
-    Ref : Paulo C. C. Freire and Norbert Wex, Mon. Not. R. Astron. Soc. 409,
-    ELL1H model parameterize the shapiro delay differently compare to ELL1
-    model. A fourier series expansion is used for the shapiro delay.
-    Ds = -2r * (a0/2 + sum(a_k*cos(k*phi)) + sum(b_k * sin(k*phi))
-    The first two harmonics are generlly absorbed by ELL1 roemer delay.
-    Thus, when ELL1 parameterize shapiro delay uses the series from the third
-    harmonic or higher.
+    Based on Freire and Wex (2010)
+
+    The :class:`~pint.models.binary_ell1.BinaryELL1H` model parameterizes the Shapiro
+    delay differently compare to the :class:`~pint.models.binary_ell1.BinaryELL1`
+    model. A fourier series expansion is used for the Shapiro delay:
+
+    .. math::
+
+        \\Delta_S = -2r \\left( \\frac{a_0}{2} + \\Sum_k (a_k \\cos k\\phi + b_k \\sin k \phi) \\right)
+
+    The first two harmonics are generlly absorbed by the ELL1 Roemer delay.
+    Thus, :class:`~pint.models.binary_ell1.BinaryELL1H` uses the series from the third
+    harmonic and higher.
+
+    References
+    ----------
+    - Freire and Wex (2010), MNRAS, 409, 199 [1]_
+
+    .. [1] https://ui.adsabs.harvard.edu/abs/2010MNRAS.409..199F/abstract
+
     """
 
     def __init__(self):
@@ -89,20 +103,14 @@ class ELL1Hmodel(ELL1BaseModel):
         """Select the fourier basis and part of the coefficent depend on the parity of the harmonics k"""
         if k % 2 == 0:
             pwr = (k + 2) / 2
-            if not derivative:
-                basis_func = np.cos
-            else:
-                basis_func = lambda x: -1 * np.sin(x)
+            basis_func = (lambda x: -1 * np.sin(x)) if derivative else np.cos
         else:
             pwr = (k + 1) / 2
-            if not derivative:
-                basis_func = np.sin
-            else:
-                basis_func = np.cos
+            basis_func = np.cos if derivative else np.sin
         return pwr, basis_func
 
     def fourier_component(self, stigma, k, factor_out_power=0):
-        """(P. Freire and N. Wex 2010) paper Eq (13)
+        """Freire and Wex (2010), Eq (13)
 
         Parameters
         ----------
@@ -118,7 +126,8 @@ class ELL1Hmodel(ELL1BaseModel):
 
         Returns
         -------
-        The coefficent of fourier component and the basis.
+        float
+            The coefficent of fourier component and the basis.
         """
 
         if k != 0:
@@ -170,9 +179,7 @@ class ELL1Hmodel(ELL1BaseModel):
             return 0, basis_func
 
     def d_STIGMA_d_H4(self):
-        if self.H3 == 0.0:
-            return 0.0
-        return 1.0 / self.H3
+        return 0.0 if self.H3 == 0.0 else 1.0 / self.H3
 
     def d_STIGMA_d_H3(self):
         if set(self.fit_params) == set(["H3", "H4"]):
@@ -196,9 +203,8 @@ class ELL1Hmodel(ELL1BaseModel):
     ):
         """Fourier series harms of shapiro delay.
 
-        One can select the start term and end term, in other
-        words, a part the fourier series term can be selected.
-        (P. Freire and N. Wex 2010) paper Eq (10)
+        One can select the start term and end term.
+        Freire and Wex (2010), Eq. (10)
 
         Parameters
         ----------
@@ -210,16 +216,15 @@ class ELL1Hmodel(ELL1BaseModel):
 
         Returns
         -------
-        The summation of harmonics
+        np.ndarray
+            The summation of harmonics
         """
         harms = np.zeros((len(selected_harms), len(phi)))
         # To prevent factor out zeros
-        if stigma == 0.0:
-            if selected_harms.min() < factor_out_power:
-                raise ValueError(
-                    "Can not factor_out_power can not bigger than"
-                    " the selected_harms."
-                )
+        if stigma == 0.0 and selected_harms.min() < factor_out_power:
+            raise ValueError(
+                "Can not factor_out_power can not bigger than" " the selected_harms."
+            )
         for ii, k in enumerate(selected_harms):
             coeff, basis_func = self.fourier_component(
                 stigma, k, factor_out_power=factor_out_power
@@ -232,29 +237,27 @@ class ELL1Hmodel(ELL1BaseModel):
     ):
         """This is a overall derivative  function."""
         # Find the right derivative  function for fourier components
-        df_name = "d_fourier_component_d_" + par.lower()
+        df_name = f"d_fourier_component_d_{par.lower()}"
         par_obj = getattr(self, par)
         try:
             df_func = getattr(self, df_name)
-        except:
+        except AttributeError:
             return 0.0 * u.Unit(None) / par_obj.Unit
         d_harms = np.zeros((len(selected_harms), len(phi)))
         # To prevent factor out zeros
-        if stigma == 0.0:
-            if selected_harms.min() < factor_out_power:
-                raise ValueError(
-                    "Can not factor_out_power can not bigger than"
-                    " the selected_harms."
-                )
+        if stigma == 0.0 and selected_harms.min() < factor_out_power:
+            raise ValueError(
+                "Can not factor_out_power can not bigger than" " the selected_harms."
+            )
         for ii, k in enumerate(selected_harms):
             coeff, basis_func = df_func(stigma, k, factor_out_power=factor_out_power)
             d_harms[ii] = coeff * basis_func(k * phi)
         return np.sum(d_harms, axis=0)
 
     def delayS3p_H3_STIGMA_approximate(self, H3, stigma, end_harm=6):
-        """Shapiro delay third harmonics or higher harms
+        """Shapiro delay using third or higher harmonics, appropriate for medium inclinations.
 
-        defined in the (P. Freire and N. Wex 2010) paper Eq (19).
+        defined in Freire and Wex (2010), Eq (19).
         """
         Phi = self.Phi()
         selected_harms = np.arange(3, end_harm + 1)
@@ -262,11 +265,10 @@ class ELL1Hmodel(ELL1BaseModel):
             selected_harms, Phi, stigma, factor_out_power=3
         )
         # There is a factor of 2 in the fharms
-        ds = -2.0 * H3 * sum_fharms
-        return ds
+        return -2.0 * H3 * sum_fharms
 
     def d_delayS3p_H3_STIGMA_approximate_d_H3(self, H3, stigma, end_harm=6):
-        """derivative  of delayS3p_H3_STIGMA with respect to H3"""
+        """derivative of delayS3p_H3_STIGMA with respect to H3"""
         Phi = self.Phi()
         selected_harms = np.arange(3, end_harm + 1)
         sum_fharms = self.ELL1H_shapiro_delay_fourier_harms(
@@ -275,7 +277,7 @@ class ELL1Hmodel(ELL1BaseModel):
         return -2.0 * sum_fharms
 
     def d_delayS3p_H3_STIGMA_approximate_d_STIGMA(self, H3, stigma, end_harm=6):
-        """derivative  of delayS3p_H3_STIGMA with respect to STIGMA"""
+        """derivative of delayS3p_H3_STIGMA with respect to STIGMA"""
         Phi = self.Phi()
         selected_harms = np.arange(3, end_harm + 1)
         sum_d_fharms = self.d_ELL1H_fourier_harms_d_par(
@@ -284,7 +286,7 @@ class ELL1Hmodel(ELL1BaseModel):
         return -2.0 * H3 * sum_d_fharms
 
     def d_delayS3p_H3_STIGMA_approximate_d_Phi(self, H3, stigma, end_harm=6):
-        """derivative  of delayS3p_H3_STIGMA with respect to Phi"""
+        """derivative of delayS3p_H3_STIGMA with respect to Phi"""
         Phi = self.Phi()
         selected_harms = np.arange(3, end_harm + 1)
         sum_d_fharms = self.d_ELL1H_fourier_harms_d_par(
@@ -293,13 +295,13 @@ class ELL1Hmodel(ELL1BaseModel):
         return -2.0 * H3 * sum_d_fharms
 
     def delayS3p_H3_STIGMA_exact(self, H3, stigma, end_harm=None):
-        """Shapiro delay third harmonics or higher harms
+        """Shapiro delay (3rd hamonic and higher) using the exact form for very high inclinations.
 
-        exact format defined in the P. Freire and N. Wex 2010 paper Eq (28).
+        Defined in Freire and Wex (2010), Eq (28).
         """
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        ds = (
+        return (
             -2
             * H3
             / stigma**3
@@ -309,16 +311,12 @@ class ELL1Hmodel(ELL1BaseModel):
                 - stigma * stigma * np.cos(2 * Phi)
             )
         )
-        return ds
 
     def d_delayS3p_H3_STIGMA_exact_d_H3(self, H3, stigma, end_harm=None):
-        """derivative  of Shapiro delay third harmonics or higher harms
-
-        exact format with respect to H3
-        """
+        """derivative of exact Shapiro delay (3rd hamonic and higher) with respect to H3"""
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_h3 = (
+        return (
             -2
             / stigma**3
             * (
@@ -327,16 +325,12 @@ class ELL1Hmodel(ELL1BaseModel):
                 - stigma * stigma * np.cos(2 * Phi)
             )
         )
-        return d_ds_d_h3
 
     def d_delayS3p_H3_STIGMA_exact_d_STIGMA(self, H3, stigma, end_harm=None):
-        """derivative  of Shapiro delay third harmonics or higher harms
-
-        exact format with respect to STIGMA
-        """
+        """derivative of exact Shapiro delay (3rd hamonic and higher) with respect to STIGMA"""
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_stigma = (
+        return (
             -2
             * H3
             / stigma**4
@@ -347,52 +341,46 @@ class ELL1Hmodel(ELL1BaseModel):
                 + stigma**2 * np.cos(Phi)
             )
         )
-        return d_ds_d_stigma
 
     def d_delayS3p_H3_STIGMA_exact_d_Phi(self, H3, stigma, end_harm=None):
-        """derivative  of Shapiro delay third harmonics or higher harms
-
-        exact format with respect to STIGMA
-        """
+        """derivative  of exact Shapiro delay (3rd hamonic and higher) with respect to phase"""
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_Phi = (
+        return (
             -4
             * H3
             / stigma**2
             * (-np.cos(Phi) / lognum + np.cos(Phi) + stigma * np.sin(2 * Phi))
         )
-        return d_ds_d_Phi
 
     def delayS_H3_STIGMA_exact(self, H3, stigma, end_harm=None):
-        """P. Freire and N. Wex 2010 paper Eq (29)"""
+        """Shapiro delay (including all harmonics) using the exact form for very high inclinations.
+
+        Defined in Freire and Wex (2010), Eq (29).
+        """
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        ds = -2 * H3 / stigma**3 * np.log(lognum)
-        return ds
+        return -2 * H3 / stigma**3 * np.log(lognum)
 
     def d_delayS_H3_STIGMA_exact_d_H3(self, H3, stigma, end_harm=None):
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_h3 = -2 / stigma**3 * np.log(lognum)
-        return d_ds_d_h3
+        return -2 / stigma**3 * np.log(lognum)
 
     def d_delayS_H3_STIGMA_exact_d_STIGMA(self, H3, stigma, end_harm=None):
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_stigma = (
+        return (
             -2
             * H3
             / stigma**4
             * (-3 * np.log(lognum) + 2 * stigma * (stigma - np.sin(Phi)) / lognum)
         )
-        return d_ds_d_stigma
 
     def d_delayS_H3_STIGMA_exact_d_Phi(self, H3, stigma, end_harm=None):
         Phi = self.Phi()
         lognum = 1 + stigma**2 - 2 * stigma * np.sin(Phi)
-        d_ds_d_Phi = 4 * H3 / stigma**2 * (np.cos(Phi) / lognum)
-        return d_ds_d_Phi
+        return 4 * H3 / stigma**2 * (np.cos(Phi) / lognum)
 
     def d_delayS_d_par(self, par):
         if set(self.fit_params) == set(["H3", "H4"]):
@@ -409,14 +397,13 @@ class ELL1Hmodel(ELL1BaseModel):
             stigma = 0.0
         else:
             raise NotImplementedError(
-                "ELL1H did not implemented %s parameter"
-                " set yet." % str(self.fit_params)
+                f"ELL1H fit not implemented for {self.fit_params} parameters"
             )
 
-        d_ds_func_name_base = "d_" + self.ds_func.__name__ + "_d_"
-        d_delayS_d_H3_func = getattr(self, d_ds_func_name_base + "H3")
-        d_delayS_d_Phi_func = getattr(self, d_ds_func_name_base + "Phi")
-        d_delayS_d_STIGMA_func = getattr(self, d_ds_func_name_base + "STIGMA")
+        d_ds_func_name_base = f"d_{self.ds_func.__name__}_d_"
+        d_delayS_d_H3_func = getattr(self, f"{d_ds_func_name_base}H3")
+        d_delayS_d_Phi_func = getattr(self, f"{d_ds_func_name_base}Phi")
+        d_delayS_d_STIGMA_func = getattr(self, f"{d_ds_func_name_base}STIGMA")
 
         d_delayS_d_H3 = d_delayS_d_H3_func(self.H3, stigma, self.NHARMS)
         d_delayS_d_Phi = d_delayS_d_Phi_func(self.H3, stigma, self.NHARMS)