+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project, at least loosely, adheres to [Semantic Versioning](https://sem
 - Methods to compute dispersion slope and to convert DM using the CODATA value of DMconst
 - `TimingModel.total_dispersion_slope` method
 - Explicit discussion of DT92 convention to DDK model
+- DDS and DDGR models
+- Module for converting between binary models
 ### Fixed
 ### Removed